Move RWS to a modern client side infrastructure
In particular, we use yarn for managing dependencies, typescript
instead of javascript, babel for transpiling to compatible JS, webpack
to handle the workflows.

The workflow idea is that all required outputs will end up in the
dist/ directory, including the main html file, minified CSS files in
css/, the JS bundles in js/, and all non-inlined images in img/.

All files are named with a hash of their content, so that the browser
can cache everything infinitely (if the content changes, the filename
changes). JS bundles are split in 3: webpack runtime, vendor code, and
our code (in theory, vendor code is more stable and could be cached
for longer, also they can be loaded in parallel). An exception is
images, because webpack can't inform the python code where we use
logo.png and friends of their new name (in the future, we could handle
the filename logo.png entirely on the client side to avoid hardcoding
it in the python code).

Configuration files expected to be shared with other frontends are
going to be in the repository root, and linked in each frontend root
(only tsconfig for the moment).

Results: from 23 requests to 14, saved 10% of bytes transmitted.